requirements.txt
kafka-python
faker
requests
psycopg2-binary
